Logical replication using stored procedure (lua-replicant)
Uses Lua consumer mechanism to emit events
New trigger method:rep providing:
  * default consumer code
  * one event per transaction (json encoded)
New insert stmt to specify genid to use
Mechanism to get/put replication seqno from SQL
Sample cdb2rcv to keep comdb2 destination in sync

TODO:
  * cleanup thread local foo; move into javasp hndl

+++ sqlite/comdb2vdbe.h
@@ -11,10 +11,10 @@ int initOpFuncDefault(OpFunc* o);
 void freeOpFunc(OpFunc* o);
 int moreRecords(OpFunc* opf);
 char* nextString(OpFunc* opf, Mem* mem);
-int nextInteger(OpFunc* opf);
+i64 nextInteger(OpFunc* opf);
 double nextReal(OpFunc* opf);
 int opFuncPrintf(OpFunc* arg, char* format, ...);
-int opFuncWriteInteger(OpFunc* opf, int val);
+int opFuncWriteInteger(OpFunc* opf, i64 val);
 int opFuncWriteReal(OpFunc* opf, double val);
 int comdb2VdbeAddExecFunc(Vdbe *p, Parse *pParse, void* arg, 
     vdbeFunc func, vdbeFuncArgFree dest);
